edit permission for user task

+++ server/controller/user_task/deleteUserTask.js
@@ -1,12 +1,14 @@
 const { deleteUserTaskQuery } = require('../../database/queries');
-// const { boomify } = require('../../utils');
+const { boomify } = require('../../utils');
 
 const deleteUserTask = async (req, res, next) => {
   try {
     const { userId, taskId } = req.params;
-    // if (+userId !== req.user.id) {
-    //   throw boomify(401, 'you dont have permission to delete user task!');
-    // }
+    const { delete: deleteTaskPermission } = req.permission;
+
+    if (!deleteTaskPermission && +userId !== req.user.id) {
+      throw boomify(401, 'you dont have permission to delete user task!');
+    }
     await deleteUserTaskQuery({ userId, taskId });
     res.status(200).json({
       statusCode: 200,
